Added Deco Item Type Fixed Full screen tab out Updated debug label in window mode

+++ Client/MirObjects/DecoObject.cs
@@ -21,7 +21,7 @@ namespace Client.MirObjects
             get { return false; }
         }
 
-        public ushort Image;
+        public int Image;
 
         public DecoObject(uint objectID)
             : base(objectID)
